+++ Backend/src/routes/CRUD.routes.js
@@ -2,14 +2,8 @@ const express = require("express");
 const Joi = require("joi");
 const appRoute = express.Router();
 const { check, validationResult } = require("express-validator");
-const {
-  delById,
-  getData,
-  getDataById,
-  search,
-  patchDataById,
-  upload,
-} = require("../transaction");
+const {handleDelete,handleGetData,handleGetDataById,handlePatchData,handleDataUpload}=require('../controller')
+
 
 const dataSchema = [
   check("Title")
@@ -38,107 +32,26 @@ const dataSchema = [
 
 //delete by id
 
-appRoute.delete("/delete", async (req, res) => {
-  try {
-    const id = req.query.id;
-    const page=req.query.page
-
-    let data = await delById({ id, req,page });
-
-    if (data.msg) {
-      console.log(data.Data);
-      return res.status(200).send({ message: data.message, data: data.Data,"Page":page });
-    }
-    console.log({ message: data.message });
-    res.status(401).send({ message: data.message });
-  } catch (error) {
-    res.status(404).send({ msg: error.message });
-  }
-});
+appRoute.delete("/delete",handleDelete );
 
 //get data
 
-appRoute.get("/data", async (req, res) => {
-  try {
-    const pageoffset = req.query.pageoffset;
-    const pageSize = req.query.pageSize;
-    const { user } = req.body;
-    const arr = await getData({ user, pageoffset, pageSize });
-    if (!arr) {
-      return res
-        .status(404)
-        .send({ msg: "fetched data Sucessfully", Data: [error.message] });
-    }
-    res.status(200).send(arr);
-  } catch (error) {
-    res.status(404).send({ msg: error.message });
-  }
-});
+appRoute.get("/data",handleGetData );
 
 // get data by id
 
-appRoute.get("/data/:id", async (req, res) => {
-  try {
-    const id = req.params.id;
-    const Data = await getDataById(id);
-    if (!Data) {
-      return res
-        .status(404)
-        .send({
-          msg: "error in fetching data Sucessfully",
-          Data: "no data found",
-        });
-    }
-    res.status(200).send({ msg: "fetched data Sucessfully", Data: Data });
-  } catch (error) {
-    res.status(404).send({ msg: error.message });
-  }
-});
+appRoute.get("/data/:id",handleGetDataById );
 
 // code for search functionality using query
 
-appRoute.get("/Search", async (req, res) => {
-  const title = req.query.search;
-  try {
-    const arr = await search(title);
-    res.status(200).send(arr);
-  } catch (error) {
-    res.status(404).send({ msg: error.message });
-  }
-});
+appRoute.get("/Search",);
 
 //update by id
 
-appRoute.patch("/update/:id", dataSchema, async (req, res) => {
-  try {
-    const id = req.params.id;
-    if (id) {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-      }
-      const data1 = req.body;
-      const data = await patchDataById({ id, data1 });
-      return res.status(200).send({ msg: "updated succesfully", Data: data });
-    }
-  } catch (error) {
-    res.status(404).send({ msg: error.message });
-  }
-});
+appRoute.patch("/update/:id",handlePatchData );
 
 //code for upload process
 
-appRoute.post("/upload", dataSchema, async (req, res) => {
-  try {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-    const data = await upload(req.body);
-    res.status(200).send({ msg: "uploaded Sucessfully" });
-  } catch (error) {
-    res.status(404).send({ msg: error.message });
-  }
-});
+appRoute.post("/upload",handleDataUpload );
 
 module.exports = { appRoute };
