+++ frontend/bcube/src/App.js
@@ -4,19 +4,22 @@ import './App.css';
 //import Container2 from './components/Container2';
 //import Sidesection from './components/Sidesection';
 import { Component } from 'react';
-//import Upload from './components/Upload';
+
 import { BrowserRouter as Router, Route,Routes } from 'react-router-dom';
-//import Overview from './components/Overview';
 
 import { toast,ToastContainer  } from 'react-toastify';
 import { connect } from 'react-redux';
 
-import Loginform from './components/Loginform';
+
 import Dashboard from './components/Dashboard';
- import Sidebar from './components/SideComponent/Sidebar';
+ 
 import Navbar from './components/Navbar';
 import UploadForm from './components/UploadForm';
-import Overview from './components/Overview';
+
+import Sidebar from './components/Sidebar';
+import OverView from './components/OverView';
+import Loginform from './components/Loginform';
+//import LoginForm from './components/Form';
 
  //import Dashboard from './components/Dashboard';
 
@@ -28,6 +31,7 @@ constructor(props) {
 
   this.state = {
      loginForm:false,
+    
   }
 
   this.toggleLoginForm=this.toggleLoginForm.bind(this)
@@ -46,6 +50,7 @@ toggleLoginForm(){
 
     this.setState((prevState) => ({
       loginForm: !prevState.loginForm,
+     
     }));
   
  
@@ -62,12 +67,12 @@ toggleLoginForm(){
  <div className="App">
        <Navbar  toggleLoginForm={this.toggleLoginForm}/>
        <div id="cont1">
-     <Sidebar/>
+    <Sidebar/>
        
         <Routes>
               {<Route path="/Dashboard" element={<Dashboard loginForm={loginForm} toggleLoginForm={this.toggleLoginForm}/>} />}
               {<Route path="/upload" element={<UploadForm  loginForm={loginForm}  />} />}
-              { <Route path="/overview/:encryptedVideoId" element={<Overview/>}/> }
+              { <Route path="/overview/:encryptedVideoId" element={<OverView/>}/> }
        </Routes>
        {this.state.loginForm && <Loginform toggleLoginForm={this.toggleLoginForm}  />}
