+++ frontend/bcube/src/components/Form/Props.js
@@ -1,11 +1,13 @@
  import { login } from "../../Redux/Redux";
 
+
 export const mapStateToProps = (state) => {
     return {
       isError: state.loginReducer.isError,
       isLogin: state.loginReducer.isLogin,
     };
   };
+
 
 export const mapDispatchToProps = (dispatch) => {
     return {
