+++ frontend/bcube/src/components/Form/index.js
@@ -1,0 +1,6 @@
+import { connect } from "react-redux";
+import {mapDispatchToProps,mapStateToProps} from './Props';
+import { withNavigateHook } from "../Navigate";
+import {LoginForm} from './LoginForm'
+
+export default connect( mapStateToProps, mapDispatchToProps)( withNavigateHook(LoginForm));
