--- frontend/bcube/src/components/LoginForm/index.js
@@ -1,6 +1,0 @@
-import { connect } from "react-redux";
-import {mapDispatchToProps,mapStateToProps} from './Props';
-
-import LoginForm from './LoginForm'
-
-export default connect( mapStateToProps, mapDispatchToProps)(LoginForm);
