+++ frontend/bcube/src/components/OverView/index.js
@@ -1,4 +1,6 @@
 import { connect } from "react-redux";
+import {mapDispatchToProps,mapStateToProps} from './Props'
 import OverView from './OverView';
-import {mapDispatchToProps,mapStateToProps} from './Props';
+
+
 export default connect( mapStateToProps, mapDispatchToProps)(OverView);
