--- frontend/bcube/src/components/SideComponent/Props.js
@@ -1,8 +1,0 @@
-const mapStateToProps = (state) => {
-    return {
-      login: state.loginReducer.isLogin,
-    };
-  };
-
-
-  export default (mapStateToProps)
