+++ frontend/bcube/src/components/Sidebar/Props.js
@@ -1,0 +1,5 @@
+export const mapStateToProps = (state) => {
+  return {
+    login: state.loginReducer.isLogin,
+  };
+};
