+++ frontend/bcube/src/components/Sidebar/index.js
@@ -1,5 +1,7 @@
+
+import Sidebar from "./Sidebar";
 import { connect } from "react-redux";
-import { mapStateToProps } from './Props';
-import {Sidebar} from './Sidebar'
+import {mapStateToProps} from './Props';
+
 
 export default connect(mapStateToProps)(Sidebar)
